requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
importlib-resources==5.10.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==21.3
pkgutil-resolve-name==1.3.10
pycodestyle==2.9.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing-extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.12.1
